+++ libs/azure-ai/langchain_azure_ai/callbacks/tracers/__init__.py
@@ -1,15 +1,11 @@
 """Tracing capabilities for Azure AI Foundry."""
 
-from langchain_azure_ai.callbacks.tracers.azure_openai_agent_tracing import (
-    AsyncAzureOpenAITracingCallback,
-    AzureOpenAITracingCallback,
-)
 from langchain_azure_ai.callbacks.tracers.inference_tracing import (
     AzureAIInferenceTracer,
+    AsyncAzureAIInferenceTracer,
 )
 
 __all__ = [
     "AzureAIInferenceTracer",
-    "AzureOpenAITracingCallback",
-    "AsyncAzureOpenAITracingCallback",
+    "AsyncAzureAIInferenceTracer",
 ]
